Lab2 linus2

+++ src/Utility/IPositionable.java
@@ -1,9 +1,6 @@
 package Utility;
 
-
-import java.math.BigDecimal;
-
 public interface IPositionable {
-    BigDecimal getY();
-    BigDecimal getX();
+    double getY();
+    double getX();
 }
